+++ src/main/java/com/techRestore/tech/restore/user/dto/reviews/ReviewRequestDTO.java
@@ -1,5 +1,7 @@
 package com.techRestore.tech.restore.user.dto.reviews;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
@@ -11,7 +13,7 @@ import lombok.Data;
 public class ReviewRequestDTO {
     @Min(1)
     @Max(5)
-    private int rating;
+    private BigDecimal rating;
     @NotBlank
     @Size(max = 1000)
     private String comment;
